Distributed Cache Source codes
Here we have the source code of distributed cache feature of hadoop
using addCacheFiles with URI and also by using GenericOptionParcer
arguments -files which is illustrated in Hadoop Definite Guide but the
StattionMetadataParcer class was missing in book. I completed the code
so that it will help readers to execute it without any problem.